Correctifs Optimisation de code (+qualité) Catégories d'article et suppression des éléments dans département Recopie de la partie sous-commission et jury. A tester

+++ src/Controller/AdministrationController.php
@@ -25,7 +25,7 @@ class AdministrationController extends BaseController
         $periodes = [];
         foreach ($this->dataUserSession->getDiplomes() as $diplome) {
             $pers = $stagePeriodeRepository->findByDiplome($diplome,
-                $diplome->getAnneeUniversitaire() ? $diplome->getAnneeUniversitaire()->getAnnee() : date('Y'));
+                $diplome->getAnneeUniversitaire() ? $diplome->getAnneeUniversitaire()->getAnnee() : (int)date('Y'));
             foreach ($pers as $periode) {
                 $periodes[] = $periode;
             }
